kubectl: remove unneeded internalclientset dependency
